[desserbee-web] api 전체적으로 연결, 자잘한 스타일 수정(아직 진행 필요), 지도 화면 중심으로 수정 (#100)
* fix(store): authorization

* feat(map):선호도태그 api

* fix(deploy): 배포 에러

* fix(map): authorization 처리 수정

* feat(desserbee-web): 유저 로그인 여부 체크

* style(desserbee-web): 바텀시트 아이콘 위치 수정

* feat(store): api 테스트

* fix(map): 위치 권한 수정

* fix(map): 현재 위치 불러오는 메서드 분리

* feat(map): 지도 중심 가져오도록 처리

* feat(map): 새로고침 버튼 추가

* fix(map): 언마운트시 클린업 함수 수정

* feat(map): 중심 좌표와 지도 화면에 따라 마커 업데이트

* style(desserbee-web): 메뉴 이미지 모달 스타일 수정

* fix(desserbee-web): 바텀시트 위치

* style(desserbee-web): 상세 페이지 화면 너비 수정

* style(desserbee-web): 한줄 리뷰 상세 페이지 수정

* feat(review): 한줄평 api

* feat(desserbee-web): 리뷰 작성 및 조회 api

* feat(desserbee-web): storeUuid 전달

* feat(desserbee-web): storeUuid 전달

* feat(map): 가게 새로 불러오기 누를 때마다 새로 가게 마커 생성, 이전 마크 제거

* fix(map): 최대 범위 3km로 수정

* feat(desserbee-web): 한줄 리뷰 페이지 - 리뷰 조회

* fix(map): 검색 반경 4km로 수정

+++ apps/desserbee-web/src/app/[lang]/(user)/(menu)/(search)/map/_components/KakaoMap/index.tsx
@@ -32,19 +32,19 @@ import StoreAPIReopository from '@repo/infrastructures/src/repositories/storeAPI
 import { KAKAO_MAP_API_URL } from '../../_consts/map';
 import {
   type NearByStoreData,
-  type SavedListData,
+  type PreferenceData,
 } from '@repo/entity/src/store';
 
 import { LocationPermissionModal } from '../../_modals/LocationPermissionModal';
 import { PortalContext } from '@repo/ui/contexts/PortalContext';
 import { useRouter } from 'next/navigation';
 import { GeolocationPermissionError } from '@repo/usecase/src/geolocationService';
-import { nearBystores } from '../../_consts/marker';
+import { ReFetchStoreBtn } from '../ReFetchStoreBtn';
+import { calculateDistance } from '../../_utils/distance';
 
 interface KakaoMapProps {
-  userPreferences: string[];
-  preferenceCategories: string[];
-  totalSavedList: SavedListData[];
+  userPreferences: number[];
+  preferenceCategories: PreferenceData[];
 }
 
 // 서비스가 모두 초기화되었는지 확인하는 헬퍼 함수
@@ -72,6 +72,7 @@ export function KakaoMap({
     storeService: null,
   });
 
+  const [isScriptLoaded, setIsScriptLoaded] = useState(false);
   const [currentPosition, setCurrentPosition] = useState<MapPosition>({
     latitude: 0,
     longitude: 0,
@@ -84,9 +85,15 @@ export function KakaoMap({
   const [isInitialized, setIsInitialized] = useState(false);
   const [isMapLoaded, setIsMapLoaded] = useState(false);
   const [nearByStores, setNearByStores] = useState<NearByStoreData[]>([]);
+  const [mapCenter, setMapCenter] = useState<MapPosition>({
+    latitude: 0,
+    longitude: 0,
+  });
 
-  const FETCH_RADIUS_M = 5000;
-  const REFETCH_THRESHOLD_M = 5000;
+  const [isFetchRequired, setIsFetchRequired] = useState(false);
+  const [isFirstLoad, setIsFirstLoad] = useState(true);
+
+  const FETCH_RADIUS_M = 4000;
   const POSITION_UPDATE_INTERVAL = 3000;
   const lastUpdateTimeRef = useRef(0);
   const isLoadingRef = useRef(false);
@@ -107,6 +114,7 @@ export function KakaoMap({
     });
   }, [closeModal, push]);
 
+  // 각 마커 클릭 - 바텀시트 열리고, 클릭한 마커 storeId 업데이트
   const handleStoreMarkerClick = useCallback(
     (storeId: string) => {
       router.replace(`?storeId=${storeId}&bottomsheet=true`, {
@@ -116,38 +124,64 @@ export function KakaoMap({
     [router],
   );
 
-  const calculateDistance = useCallback(
-    (pos1: MapPosition, pos2: MapPosition): number => {
-      const R = 6371;
-      const dLat = ((pos2.latitude - pos1.latitude) * Math.PI) / 180;
-      const dLon = ((pos2.longitude - pos1.longitude) * Math.PI) / 180;
-      const a =
-        Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-        Math.cos((pos1.latitude * Math.PI) / 180) *
-          Math.cos((pos2.latitude * Math.PI) / 180) *
-          Math.sin(dLon / 2) *
-          Math.sin(dLon / 2);
-      const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-      return R * c;
-    },
-    [],
-  );
+  // 제일 먼저 서비스 초기화
+  const initializeServices = () => {
+    console.log('initializeServices: 서비스 초기화 시작');
 
-  const handleMoveToCurrentPosition = useCallback(() => {
-    if (servicesRef.current.mapService && isMapLoaded) {
-      servicesRef.current.mapService.setMapCenter(currentPosition);
-    }
-  }, [isMapLoaded, currentPosition]);
+    const mapService = new MapService({
+      mapController: new KakaoMapController(),
+    });
+    console.log('initializeServices: MapService 초기화 완료 ✅');
 
-  const mapPanelProps = useMemo(
-    () => ({
-      moveToCurrentPosition: handleMoveToCurrentPosition,
-    }),
-    [handleMoveToCurrentPosition],
-  );
+    const geoService = new GeolocationService({
+      geolocationController: new GeolocationController(
+        new KalmanLocationFilter(),
+        new MovingAverageFilter(3),
+      ),
+    });
+    console.log('initializeServices: GeolocationService 초기화 완료 ✅');
 
-  // 지도 로드 후
-  // 1. 현재 위치 기반으로 일정 거리 내의 새로운 가게 불러오기
+    const storeService = new StoreService({
+      storeRepository: new StoreAPIReopository(),
+    });
+    console.log('initializeServices: StoreService 초기화 완료 ✅');
+
+    return { mapService, geoService, storeService };
+  };
+
+  // 거리 계산해서 가게 업데이트 필요 판단
+  // const determineFetch = (
+  //   lastFetchPosition: MapPosition,
+  //   currentMapPosition: MapPosition,
+  // ) => {
+  //   const distanceFromLastFetch = calculateDistance(
+  //     lastFetchPosition,
+  //     currentMapPosition,
+  //   );
+
+  //   console.log(
+  //     'determineFetch: 마지막 데이터 요청 위치와의 거리 📏:',
+  //     distanceFromLastFetch,
+  //     'km',
+  //   );
+
+  //   if (
+  //     lastFetchPosition.latitude === 0 ||
+  //     distanceFromLastFetch > REFETCH_THRESHOLD_M
+  //   ) {
+  //     console.log(
+  //       'determineFetch: 재요청 할 때 됨, (임계값 초과), 주변 가게 정보 업데이트 시작 ✅',
+  //     );
+  //     setIsFetchRequired(true);
+  //   } else {
+  //     console.log(
+  //       'determineFetch: 재요청 임계값 이내, 아직 새로운 가게 재요청 안함 ⌛',
+  //     );
+  //     setIsFetchRequired(false);
+  //   }
+  // };
+
+  // 지금 지도에서 위치한 근처의 가게들 fetch
   const fetchNearbyStores = useCallback(
     async (position: MapPosition) => {
       try {
@@ -191,12 +225,12 @@ export function KakaoMap({
     [retryCount],
   );
 
-  // 2. 새로운 위치 저장 + 새로운 위치에 따른 새로운 가게들 정보 담은 클러스터 마커 추가
   const updateLastFetchPosition = useCallback((position: MapPosition) => {
     console.log('updateLastFetchPosition: 마지막 fetch 위치 저장 ✅');
     setLastFetchPosition(position);
   }, []);
 
+  // 지도 중심, 반경내 상점들 받아서 마커 및 클러스터 마커 추가
   const updateNewClusterMarkers = useCallback(
     async (position: MapPosition, stores: NearByStoreData[]) => {
       if (!areServicesInitialized(servicesRef.current)) {
@@ -207,6 +241,9 @@ export function KakaoMap({
       }
       try {
         isLoadingRef.current = true;
+
+        console.log('updateCurrentMarker: 이전 마커 모두두 제거 시작 🗑️');
+        await servicesRef.current.mapService?.clearAllMarkers();
 
         if (stores.length !== 0) {
           console.log(
@@ -235,86 +272,51 @@ export function KakaoMap({
     [servicesRef, handleStoreMarkerClick, updateLastFetchPosition],
   );
 
-  // 3. tracking 중 지속적으로 실행되는 메서드
-  const onPositionSuccess = useCallback(
+  // 실시간 마커 한개 ! 업데이트 (geo)
+  const updateCurrentMarker = useCallback(
     async (position: MapPosition) => {
       try {
         if (!areServicesInitialized(servicesRef.current)) {
           console.log(
-            'onPositionSuccess: 서비스들이 초기화되지 않음, 위치 업데이트 중지 🛑',
+            'updateCurrentMarker: 서비스들이 초기화되지 않음, 위치 업데이트 중지 🛑',
           );
           return;
         }
 
         console.log(
-          'onPositionSuccess: 트래킹 시작, 위치 업데이트 콜백 실행',
+          'updateCurrentMarker: 트래킹 시작, 위치 업데이트 콜백 실행',
           position,
         );
 
         const now = Date.now();
         if (now - lastUpdateTimeRef.current < POSITION_UPDATE_INTERVAL) {
-          console.log('onPositionSuccess: 업데이트 간격이 너무 짧음, 스킵 🛑');
+          console.log(
+            'updateCurrentMarker: 업데이트 간격이 너무 짧음, 스킵 🛑',
+          );
           return;
         }
 
         if (isLoadingRef.current) {
-          console.log('onPositionSuccess: 이전 업데이트가 진행 중, 스킵 🛑');
+          console.log('updateCurrentMarker: 이전 업데이트가 진행 중, 스킵 🛑');
           return;
         }
 
         lastUpdateTimeRef.current = now;
-        console.log('onPositionSuccess: 위치 업데이트 시작 🚩');
+        console.log('updateCurrentMarker: 위치 업데이트 시작 🚩');
 
-        console.log('onPositionSuccess: 현재 위치 마커 제거 시작 🗑️');
+        console.log('updateCurrentMarker: 현재 위치 마커 제거 시작 🗑️');
         await servicesRef.current.mapService?.removeCurrentPositionMarker();
 
-        console.log('onPositionSuccess: 새로운 현재 위치 마커 추가 📍');
+        console.log('updateCurrentMarker: 새로운 현재 위치 마커 추가 📍');
         await servicesRef.current.mapService?.addCurrentPositionMaker(
           position,
           userMarkerImage.src,
         );
         setCurrentPosition(position);
-        console.log('onPositionSuccess: 새로운 현재 위치 저장 🧍‍♂️');
-
-        const distanceFromLastFetch = calculateDistance(
-          lastFetchPosition,
-          position,
-        );
-        console.log(
-          'onPositionSuccess: 마지막 데이터 요청 위치와의 거리 📏:',
-          distanceFromLastFetch,
-          'km',
-        );
-
-        if (
-          lastFetchPosition.latitude === 0 ||
-          distanceFromLastFetch > REFETCH_THRESHOLD_M
-        ) {
-          console.log(
-            'onPositionSuccess: 재요청 할 때 됨, (임계값 초과), 주변 가게 정보 업데이트 시작 ✅',
-          );
-
-          //TODO: 가짜 데이터
-          // const stores = nearBystores;
-          const nearByStores =
-            await servicesRef.current.storeService!.getNearbyStores({
-              latitude: position.latitude,
-              longitude: position.longitude,
-              radius: FETCH_RADIUS_M,
-            });
-
-          await updateNewClusterMarkers(position, nearByStores);
-          console.log(
-            'onPositionSuccess: 주변 가게 정보 업데이트 완료, 클러스터 마커들도 새로 추가 📍',
-          );
-        } else {
-          console.log(
-            'onPositionSuccess: 재요청 임계값 이내, 아직 새로운 가게 재요청 안함 ⌛',
-          );
-        }
+        console.log('updateCurrentMarker: 새로운 현재 위치 저장 🧍‍♂️');
       } catch (error) {
         console.error(
-          'onPositionSuccess: 위치 마커 업데이트 중 오류 발생 ⚠️:',
+          'updateCurrentMarker: 위치 마커 업데이트 중 오류 발생 ⚠️:',
           error,
         );
         if (
@@ -322,47 +324,83 @@ export function KakaoMap({
           error.message === 'PERMISSION_DENIED'
         ) {
           console.log(
-            'onPositionSuccess: 위치 권한 거부됨, 권한 요청 모달 표시 🪧',
+            'updateCurrentMarker: 위치 권한 거부됨, 권한 요청 모달 표시 🪧',
           );
           openPermissionModal();
         } else {
-          console.error('onPositionSuccess: 위치 권한 외 오류 발생', error);
+          console.error('updateCurrentMarker: 위치 권한 외 오류 발생', error);
         }
         return;
       }
     },
-    [
-      calculateDistance,
-      lastFetchPosition,
-      updateNewClusterMarkers,
-      openPermissionModal,
-    ],
+    [openPermissionModal],
   );
 
-  // 0. 서비스 시작
-  const initializeServices = () => {
-    console.log('initializeServices: 서비스 초기화 시작');
-
-    const mapService = new MapService({
-      mapController: new KakaoMapController(),
+  const handleInitialGeoPositonFetch = async (
+    initializedServices: {
+      geoService: GeolocationService;
+    },
+    openPermissionModal: () => void,
+  ) => {
+    const permissionStatus = await navigator.permissions.query({
+      name: 'geolocation',
     });
-    console.log('initializeServices: MapService 초기화 완료 ✅');
+    console.log('handleInitialGeoPositonFetch: 위치 권한 확인');
 
-    const geoService = new GeolocationService({
-      geolocationController: new GeolocationController(
-        new KalmanLocationFilter(),
-        new MovingAverageFilter(3),
-      ),
+    if (permissionStatus.state === 'denied') {
+      console.log('handleInitialGeoPositonFetch: 위치 권한 거부됨');
+      openPermissionModal();
+      return null;
+    }
+
+    console.log('handleInitialGeoPositonFetch: 현재 위치 정보 요청 시작 🏃');
+    const result = await initializedServices.geoService.getCurrentPosition();
+
+    if ('errorMessage' in result) {
+      console.log('handleInitialGeoPositonFetch: 오류 발생 ⚠️:', {
+        message: result.errorMessage,
+        type: result.errorType,
+      });
+      console.log(
+        'handleInitialGeoPositonFetch: 현재 위치 권한 상태 다시 확인:',
+        {
+          state: permissionStatus.state,
+          errorType: result.errorType,
+        },
+      );
+
+      if (result.errorType === 'POSITION_UNAVAILABLE') {
+        console.log('handleInitialGeoPositonFetch: GPS 사용 불가');
+        setError('GPS를 활성화하고 다시 시도해주세요.');
+      } else if (result.errorType === 'TIMEOUT') {
+        console.log('handleInitialGeoPositonFetch: 위치 정보 요청 시간 초과');
+        setError('위치 정보를 가져오는데 시간이 너무 오래 걸립니다.');
+      } else {
+        setError('위치 정보를 가져오는데 실패했습니다.');
+      }
+      return null;
+    }
+
+    console.log('handleInitialGeoPositonFetch: 현재 위치 정보 획득 성공 🚩:', {
+      latitude: result.latitude,
+      longitude: result.longitude,
     });
-    console.log('initializeServices: GeolocationService 초기화 완료 ✅');
 
-    const storeService = new StoreService({
-      storeRepository: new StoreAPIReopository(),
-    });
-    console.log('initializeServices: StoreService 초기화 완료 ✅');
+    setCurrentPosition(result);
+    console.log('handleInitialGeoPositonFetch: 현재 위치 저장 🚩');
 
-    return { mapService, geoService, storeService };
+    return result;
   };
+
+  const handleMapCenterChange = useCallback(() => {
+    if (servicesRef.current.mapService) {
+      const center = servicesRef.current.mapService.getMapCenter();
+      setMapCenter(center);
+
+      // 거리 기반으로 fetch 필요성 판단
+      // determineFetch(lastFetchPosition, center);
+    }
+  }, []);
 
   const loadMap = async (initializedServices: {
     mapService: MapService;
@@ -377,60 +415,30 @@ export function KakaoMap({
     }
 
     try {
-      const permissionStatus = await navigator.permissions.query({
-        name: 'geolocation',
-      });
-      console.log('loadMap - geoService.getCurrentPosition(): 위치 권한 확인');
-
-      if (permissionStatus.state === 'denied') {
-        console.log('loadMap: 위치 권한 거부됨');
-        openPermissionModal();
-        return;
-      }
-
-      console.log(
-        'loadMap - geoService.getCurrentPosition(): 현재 위치 정보 요청 시작 🏃',
+      // 첫 위치는 현재 위치
+      const result = await handleInitialGeoPositonFetch(
+        initializedServices,
+        openPermissionModal,
       );
-      const result = await initializedServices.geoService.getCurrentPosition();
 
-      if ('errorMessage' in result) {
-        console.log(
-          'loadMap - geoService.getCurrentPosition(): 오류 발생 ⚠️:',
-          {
-            message: result.errorMessage,
-            type: result.errorType,
-          },
+      if (!result) {
+        console.error(
+          '현재 위치 로딩에 문제가 생겼습니다.⚠️ 지도 로드 중지 🛑',
         );
-        console.log('loadMap: 현재 위치 권한 상태 다시 확인:', {
-          state: permissionStatus.state,
-          errorType: result.errorType,
-        });
-
-        if (result.errorType === 'POSITION_UNAVAILABLE') {
-          console.log('loadMap: GPS 사용 불가');
-          setError('GPS를 활성화하고 다시 시도해주세요.');
-        } else if (result.errorType === 'TIMEOUT') {
-          console.log('loadMap: 위치 정보 요청 시간 초과');
-          setError('위치 정보를 가져오는데 시간이 너무 오래 걸립니다.');
-        } else {
-          setError('위치 정보를 가져오는데 실패했습니다.');
-        }
         return;
       }
-
-      console.log('loadMap: 현재 위치 정보 획득 성공 🚩:', {
-        latitude: result.latitude,
-        longitude: result.longitude,
-      });
-
-      setCurrentPosition(result);
-      console.log('loadMap: 현재 위치 저장 🚩');
 
       console.log('loadMap: 지도 초기화 시작');
       await initializedServices.mapService.initializeMap(
         mapRef.current,
         result,
       );
+
+      console.log('loadMap: 지도 중심 추적 시작');
+      initializedServices.mapService.addCenterChangedListener(
+        handleMapCenterChange,
+      );
+
       console.log('loadMap: 지도 초기화 완료 🗺️');
 
       console.log('loadMap: 현재 위치 마커 추가 시작');
@@ -440,24 +448,30 @@ export function KakaoMap({
       );
       console.log('loadMap: 현재 위치 마커 추가 완료 📍');
 
-      await initializedServices.mapService.setMapCenter(result);
-      console.log('loadMap: 불러온 위치로 지도 중심 위치 변경');
+      initializedServices.geoService.startWatchingPosition(
+        updateCurrentMarker,
+        {
+          enableHighAccuracy: true,
+          timeout: 5000,
+          maximumAge: 0,
+        },
+      );
+      console.log('loadMap: 실시간 위치 추적 시작');
+
+      // 첫 로드 시에만 지도 중심 위치 변경
+      if (isFirstLoad) {
+        await initializedServices.mapService.setMapCenter(result);
+        console.log('loadMap: 불러온 위치로 지도 중심 위치 변경');
+        setIsFirstLoad(false); // 첫 로드 이후로는 실행되지 않도록 설정
+      }
 
       console.log('loadMap: 주변 가게 마커 추가 시작');
-
       await initializedServices.mapService.addMarkersWithClustering(
         nearByStores,
         storeMarkerImage.src,
         handleStoreMarkerClick,
       );
       console.log('loadMap: 주변 가게 마커 추가 완료 📍');
-
-      console.log('loadMap: 실시간 위치 추적 시작');
-      initializedServices.geoService.startWatchingPosition(onPositionSuccess, {
-        enableHighAccuracy: true,
-        timeout: 5000,
-        maximumAge: 0,
-      });
     } catch (err) {
       console.error('loadMap: 지도 초기화 중 오류 발생 ⚠️:', err);
       if (err instanceof GeolocationPermissionError) {
@@ -470,16 +484,54 @@ export function KakaoMap({
     }
   };
 
+  // 현재 실시간 유저 위치로 이동
+  const handleMoveToCurrentPosition = useCallback(() => {
+    if (servicesRef.current.mapService && isMapLoaded) {
+      servicesRef.current.mapService.setMapCenter(currentPosition);
+    }
+  }, [isMapLoaded, currentPosition]);
+
+  // 버튼에 연동
+  const mapPanelProps = useMemo(
+    () => ({
+      moveToCurrentPosition: handleMoveToCurrentPosition,
+    }),
+    [handleMoveToCurrentPosition],
+  );
+
   useEffect(() => {
     return () => {
-      if (servicesRef.current.geoService && servicesRef.current.mapService) {
+      if (
+        servicesRef.current.geoService &&
+        servicesRef.current.mapService &&
+        isInitialized
+      ) {
         servicesRef.current.geoService.stopWatchingPosition();
         servicesRef.current.mapService.removeCurrentPositionMarker();
       }
     };
-  }, [servicesRef]);
+  }, [servicesRef, isInitialized]);
 
-  const [isScriptLoaded, setIsScriptLoaded] = useState(false);
+  const handleRefetchBtnClick = () => {
+    setIsFetchRequired(true);
+  };
+
+  const mapCenterRef = useRef(mapCenter);
+
+  useEffect(() => {
+    mapCenterRef.current = mapCenter;
+  }, [mapCenter]);
+
+  useEffect(() => {
+    const fetchAndUpdate = async () => {
+      const stores = await fetchNearbyStores(mapCenterRef.current);
+      if (stores) {
+        await updateNewClusterMarkers(mapCenterRef.current, stores);
+        setIsFetchRequired(false);
+      }
+    };
+    fetchAndUpdate();
+  }, [isFetchRequired, fetchNearbyStores, updateNewClusterMarkers]);
 
   return (
     <div>
@@ -554,6 +606,7 @@ export function KakaoMap({
           categories={preferenceCategories}
         />
         <MapPanel {...mapPanelProps} />
+        <ReFetchStoreBtn refetchStore={handleRefetchBtnClick} />
       </div>
     </div>
   );
